fpg/fpg/api_urls.py
from rest_framework import routers
from django.conf.urls import url, include


router = routers.DefaultRouter()

urlpatterns = [
    url(r'^', include(router.urls)),
    url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework'))
]