requirements.txt
Django==1.8
Pillow==2.8.1
argparse==1.2.1
django-debug-toolbar==1.3.0
django-extensions==1.5.2
django-js-reverse==0.5.0
django-model-utils==2.2
django-photologue==3.1.1
django-registration-redux==1.1
django-sortedm2m==0.9.4
djangorestframework==3.1.1
ipdb==0.8
ipython==3.1.0
ply==3.4
psycopg2==2.6
six==1.9.0
slimit==0.8.1
sqlparse==0.1.14
wsgiref==0.1.2
